Refactor JavaScript file structure
