+++ assets/js/mnist/mnistdraw.js
