+++ assets/js/mnist/mnistsocket.js
