+++ assets/js/nlp/nlpsocket.js
