+++ lib/bayberry_web/views/main_view.ex
@@ -5,25 +5,25 @@ defmodule BayberryWeb.MainView do
     case assigns.script_path do
       "/" ->
         ~S"""
-        <script>require("js/socket/mnistcloud").DigitCloud.run()</script>
-        <script>require("js/topics").LdaTopics.run()</script>
-        <script>require("js/wordcloud").WordCloud.run()</script>
-        <script>require("js/socket/visitors").VisitorMap.run()</script>
+        <script>require("js/main/digits").Digits.run()</script>
+        <script>require("js/main/topics").Topics.run()</script>
+        <script>require("js/main/visitors").Visitors.run()</script>
+        <script>require("js/main/wordcloud").WordCloud.run()</script>
         """
         |> raw
       "/mnist" ->
         ~S"""
-        <script>require("js/socket/mnistsocket").MnistSocket.run()</script>
+        <script>require("js/mnist/mnistsocket").MnistSocket.run()</script>
         """
         |> raw
       "/nlp" ->
         ~S"""
-        <script>require("js/socket/nlpsocket").NlpSocket.run()</script>
+        <script>require("js/nlp/nlpsocket").NlpSocket.run()</script>
         """
         |> raw
       "/twitter" ->
         ~S"""
-        <script>require("js/socket/twittersocket").TwitterSocket.run()</script>
+        <script>require("js/twitter/twittersocket").TwitterSocket.run()</script>
         """
         |> raw
       "/blank" ->
